+++ src/pages/auth/hooks/useEmailPrompt.ts
@@ -1,13 +1,11 @@
 import { yupResolver } from '@hookform/resolvers/yup'
-import { atom, useAtomValue, useSetAtom } from 'jotai'
-import { atomWithStorage } from 'jotai/utils'
-import { useEffect, useState } from 'react'
+import { useAtomValue, useSetAtom } from 'jotai'
 import { useForm } from 'react-hook-form'
 import * as yup from 'yup'
 
 import { useAuthContext } from '../../../context'
 import { useUserEmailUpdateMutation } from '../../../types'
-import { authUserAtom } from '../state'
+import { dontAskAgainAtom, shouldPromptAtom } from '../state/emailPromptAtom'
 
 type FormValues = Record<string, any>
 
@@ -23,17 +21,8 @@ const emailSchema = yup.object().shape({
     }),
 })
 
-export const dontAskAgainAtom = atomWithStorage('dontAskAgain', false)
-
-export const shouldPromptAtom = atom((get) => {
-  const user = get(authUserAtom)
-  const dontAskAgain = get(dontAskAgainAtom)
-  return user.id && !user.email && !dontAskAgain
-})
-
 export const useEmailPrompt = () => {
   const { user, setUser } = useAuthContext()
-  const [enableSave, setEnableSave] = useState(false)
   const setDontAskAgain = useSetAtom(dontAskAgainAtom)
   const shouldPrompt = useAtomValue(shouldPromptAtom)
   const [updateUserEmail] = useUserEmailUpdateMutation({
@@ -58,9 +47,7 @@ export const useEmailPrompt = () => {
     },
   })
 
-  useEffect(() => {
-    setEnableSave(isValid && isDirty)
-  }, [isValid, isDirty])
+  const enableSave = isValid && isDirty
 
   const onSubmit = (data: FormValues) => {
     if (data.email) {
